Refactor Quicklinks to functional component with logic hook and state
