+++ Pzl.Part.QuickLinks/src/webparts/quickLinks/components/QuickLinks.module.scss
@@ -30,7 +30,6 @@
     }
     .linkCategoryHeading {
       @include ms-fontSize-xl;
-      //color: "[theme: themePrimary, default: #2C7399]";
       width: 100%;
       white-space: nowrap;
     }
@@ -44,17 +43,14 @@
         display: inline-block;
         margin: 0 30px 0 0;
         margin-bottom: 5px;
-        //color: $ms-color-neutralPrimary;
         span,
         i {
           display: inline-block;
-          //color: $ms-color-neutralPrimary;
           vertical-align: middle;
         }
         span {
           @include ms-fontWeight-regular;
           @include ms-fontSize-mPlus;
-          //color: $ms-color-neutralPrimary;
           text-overflow: ellipsis;
           width: 200px;
           white-space: nowrap;
