+++ Pzl.Part.QuickLinks/src/webparts/quickLinks/components/index.ts
@@ -1,0 +1,2 @@
+export * from './QuickLinks'
+export * from './types'
