+++ Pzl.Part.QuickLinks/src/webparts/quickLinks/components/types.ts
@@ -14,3 +14,22 @@ export interface IQuickLinksProps {
   webServerRelativeUrl: string
   linkClickWebHook: string
 }
+
+
+export interface IQuickLinksState {
+  linkStructure: Array<ICategory>
+}
+
+export interface ILink {
+  id?: number
+  displayText: string
+  url: string
+  icon: string
+  category: string
+  priority: string
+  openInSameTab: boolean
+}
+export interface ICategory {
+  links: Array<ILink>
+  displayText: string
+}
