+++ Pzl.Part.QuickLinks/src/webparts/quickLinks/components/useQuickLinksState.ts
@@ -1,0 +1,17 @@
+import { useState } from 'react'
+import { IQuickLinksState } from './types'
+
+/**
+ * Hook for QuickLinks component state
+ */
+export const useQuickLinksState = () => {
+  const [state, $setState] = useState<IQuickLinksState>({
+    linkStructure: []
+  })
+
+  const setState = (newState: Partial<IQuickLinksState>) => {
+    $setState((_state) => ({ ..._state, ...newState }))
+  }
+
+  return { state, setState }
+}
